% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.mean.R
\name{col.mean}
\alias{col.mean}
\title{Give the hex color weighted average of two hex colors}
\usage{
col.mean(col1, col2, w1 = 0.5, w2 = 0.5, direct = TRUE)
}
\arguments{
\item{col1}{first hex color}

\item{w1}{weight to give first hex color (w1 + w2 = 1)}

\item{w2}{weight to give second hex color (w1 + w2 = 1)}

\item{direct}{if TRUE, will get the rgb average. If FALSE will give the hsv average}

\item{col1}{second hex color}
}
\value{
new hex color
}
\description{
\code{col.mean()} takes two hex colors and two sets of weights to give the average color
The direct method gives the rgb average, whereas the indirect method gives the hsv average
by finding the shortest angle needed to be traversed by the hue parameter
}
