% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.shape.R
\name{color.shape}
\alias{color.shape}
\title{Give the hex color to a shape based on hsv coordinates and additional rules}
\usage{
color.shape(ang, s = 1, v = 1, rot = 0, type = "kite")
}
\arguments{
\item{ang}{numeric angle in degrees}

\item{s}{numeric saturation for color [0,1]}

\item{v}{numeric value for color [0,1]}

\item{rot}{numeric angle in degrees for rotating the hue parameter}

\item{type}{either "kite" or "dart" for some additional arbitrary color rules}
}
\value{
list of numeric coordinates for the new shapes from the transform
}
\description{
\code{color.shape()} takes hsv coordinates and returns the rgb hexcode, 
but with some additional rules, for artistic purposes
}
