% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvcol.R
\name{hsvcol}
\alias{hsvcol}
\title{Give the hex color to a shape based on hsv coordinates with additional h parameters}
\usage{
hsvcol(ang, s = 1, v = 1, rot = 0)
}
\arguments{
\item{ang}{numeric angle in degrees}

\item{s}{numeric saturation for color [0,1]}

\item{v}{numeric value for color [0,1]}

\item{rot}{numeric angle in degrees for rotating the hue parameter}

\item{type}{either "kite" or "dart" for some additional arbitrary color rules}
}
\value{
list of numeric coordinates for the new shapes from the transform
}
\description{
\code{hsvcol()} takes hsv coordinates and returns the rgb hexcode, 
in addition to rotating hue by \code{rot}
Requires "ColorPalette" for hsv2rgb()
}
