% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penrose.R
\docType{package}
\name{penrose}
\alias{penrose}
\alias{penrose-package}
\title{penrose}
\description{
This package contains a few functions to create a P2 penrose tiling by deflation of kites and darts. 
The main function for plotting the figure is penrose.full().
}
