% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate and scale a point from a reference point}
\usage{
rotate(x, y, ax.x, ax.y, rot = 0, sca = 1)
}
\arguments{
\item{x}{x coordinate of ending point}

\item{y}{y coordinate of ending point}

\item{ax.x}{x coordinate of starting point}

\item{ax.y}{y coordinate of ending point}

\item{rot}{numeric degrees to rotate the ending point}

\item{sca}{numeric amount to scale the distance from starting point}
}
\value{
returns the new x and y point after rotation/scaling
}
\description{
\code{rotate()} returns the new point  from one point to another with modifications
}
