% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.set.R
\name{rotate.set}
\alias{rotate.set}
\title{Rotate and scale a set of points from a reference point}
\usage{
rotate.set(xs, ys, ax.x, ax.y, rot = 0, sca = 1)
}
\arguments{
\item{ax.x}{x coordinates of starting point}

\item{ax.y}{y coordinate of ending point}

\item{rot}{numeric degrees to rotate the ending point}

\item{sca}{numeric amount to scale the distance from starting point}

\item{x}{x coordinates of ending points}

\item{y}{y coordinates of ending points}
}
\value{
returns the new x and y point after rotation/scaling
}
\description{
\code{rotate.set()} returns the angle from one point to another with modifications
}
