% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.angle.R
\name{simple.angle}
\alias{simple.angle}
\title{Give angle between two points}
\usage{
simple.angle(x, y, ax.x, ax.y)
}
\arguments{
\item{x}{x coordinate of ending point}

\item{y}{y coordinate of ending point}

\item{ax.x}{x coordinate of starting point}

\item{ax.y}{y coordinate of ending point}
}
\value{
the angle in degrees
}
\description{
\code{simple.angle()} returns the angle from one point to another
}
