% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.dist.R
\name{simple.dist}
\alias{simple.dist}
\title{Give 2D Euclidean distance between two points}
\usage{
simple.dist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x coordinate of first point}

\item{y1}{y coordinate of first point}

\item{x2}{x coordinate of second point}

\item{y2}{y coordinate of the second point}
}
\value{
the distance
}
\description{
\code{simple.dist()} returns the simple distance between two points
}
