% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.shape.R
\name{split.shape}
\alias{split.shape}
\title{Split a kite or dart according to penrose construction by deflation}
\usage{
\method{split}{shape}(xs, ys, type, core.x, core.y)
}
\arguments{
\item{xs}{numeric vector containing the x coordinates of the points of the shape to be split.}

\item{ys}{numeric vector containing the y coordinates of the points of the shape to be split.}

\item{type}{either "kite" or "dart" to specify the shape}

\item{core.x}{numeric x coordinate of true center of the penrose}

\item{core.y}{numeric y coordinate of true center of the penrose}
}
\value{
list of sets of numeric coordinates for the new shapes from the split
}
\description{
\code{split.shape()} takes in x coordinates and y coordinates of a point, as well as the 
penrose's true center. You must also specify whether it is a kite or dart.
}
