% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanh.tween.R
\name{tanh.tween}
\alias{tanh.tween}
\title{Create tweened intervals for animation frames based on arctangents}
\usage{
\method{tanh}{tween}(min = 0, max = 1, n.interval = 5, scale = 5)
}
\arguments{
\item{min}{integer value for first "state"}

\item{max}{integer value for last "state"}

\item{n.interval}{integer value for number of frames between each state}

\item{scale}{numeric parameter for scaling the arctangent function and severity of tweening}
}
\value{
vector of intervals for tweening. 
Final number of points will be: ( n.interval * 2 - 1 ) * ( max - min )
}
\description{
\code{tanh.tween()} creates tweening intervals stopping at integers
}
