% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.shape.R
\name{transform.shape}
\alias{transform.shape}
\title{Rotate and scale a shape according to parameters}
\usage{
\method{transform}{shape}(xs, ys, type, core.x = 0, core.y = 0,
  core.rot = 0, core.sca = 1, cent.rot = 0, cent.sca = 1, ticker = 0)
}
\arguments{
\item{xs}{numeric vector containing the x coordinates of the points of the shape to be transformed.}

\item{ys}{numeric vector containing the y coordinates of the points of the shape to be transformed.}

\item{core.x}{numeric x coordinate of true center of the penrose}

\item{core.y}{numeric y coordinate of true center of the penrose}

\item{core.rot}{numeric degrees to rotate the set of all points by, hinged on the true center}

\item{core.sca}{numeric amount to scale the set of all points by, relative to the true center}

\item{cent.rot}{numeric degrees to rotate the set of all points by, hinged on the center of points}

\item{cent.sca}{numeric amount to scale the set of all points by, relative to the center of points}

\item{ticker}{additional outside parameter for miscellaneous use}
}
\value{
list of numeric coordinates for the new shapes from the transform
}
\description{
\code{transform.shape()} takes in x coordinates and y coordinates of a point, as well as the 
penrose's true center. Additional parameters are used to rotate and scale the data.
}
